+++ src/Symfony/Component/Yaml/CHANGELOG.md
@@ -4,6 +4,12 @@ CHANGELOG
 3.1.0
 -----
 
+ * Added support for customizing the YAML parser behavior through an optional bit field:
+
+   ```php
+   Yaml::parse('{ "foo": "bar", "fiz": "cat" }', Yaml::EXCEPTION_ON_INVALID_TYPE | Yaml::PARSE_OBJECT | Yaml::PARSE_OBJECT_FOR_MAP);
+   ```
+
  * Added support for customizing the dumped YAML string through an optional bit field:
 
    ```php